Extract B32 to separate file
